Fix numeric cursor state value comparison for boundary check

+++ airbyte-integrations/connectors/source-mssql/src/main/kotlin/io/airbyte/integrations/source/mssql/MsSqlServerJdbcPartition.kt
@@ -59,7 +59,7 @@ fun stateValueToJsonNode(field: Field, stateValue: String?): JsonNode {
                     Jsons.valueToTree(stateValue?.takeIf { it.isNotEmpty() }?.toBigInteger())
                 }
                 LeafAirbyteSchemaType.NUMBER -> {
-                    Jsons.valueToTree(stateValue?.takeIf { it.isNotEmpty() }?.toDouble())
+                    Jsons.valueToTree(stateValue?.takeIf { it.isNotEmpty() }?.toBigDecimal())
                 }
                 LeafAirbyteSchemaType.BINARY -> {
                     val ba = Base64.getDecoder().decode(stateValue!!)
